intent file with Intent class moved to be alongside Ability class in ability.py file.
