+++ pyttman/core/intent.py
