--- tests/integration/test_routing.py
@@ -1,19 +1,0 @@
-from unittest import TestCase
-import pyttman
-import testing_project.settings as settings
-from tests.mockups import ContactsAbility
-
-pyttman.load_settings(settings)
-
-from pyttman.core.communication.models.containers import Message
-from pyttman.core.parsing.routing import FirstMatchingRouter
-
-
-class TestLinearSearchFirstMatchingRouter(TestCase):
-    def test_reply(self):
-        update_contact_message = Message("register 1112222442 as new number for John")
-        contact_feature = ContactsAbility()
-        router = FirstMatchingRouter()
-        router.abilities = (contact_feature,)
-        self.assertEqual(router.get_reply(message=update_contact_message).as_list(),
-                         ['Updated', "John's", 'number', 'to', '1112222442'])
